Implement a power light sampler.
The problem is in the spaceship scene; there is a red light source in
the cockpit consisting of ~1200 triangles (aka tiny light sources) that
get sampled almost all the time because the second light is just a quad
(two big triangles) above the spaceship.

A BVH-based light sampler as PBRT uses is might be even better but for
now implement the simplest working solution: Sampling according to the
emitted power of a light source. This way the big light source above the
ship gets sampled much more often and the image looks much better.